+++ tests/testthat/test-lazy-ops.R
@@ -103,7 +103,7 @@ test_that("arranges captures DESC", {
   out <- lazy_frame(x = 1:3, y = 3:1) %>%
     arrange(desc(x))
 
-  expect_equal(op_sort(out)[[1]], quote(desc(x)))
+  expect_equal(op_sort(out), list(quote(desc(x))))
 })
 
 test_that("multiple arranges combine", {
@@ -111,8 +111,15 @@ test_that("multiple arranges combine", {
     arrange(x) %>%
     arrange(y)
 
-  vars <- op_sort(out)
-  expect_length(vars, 2)
-  expect_equal(vars[[1]], quote(x))
-  expect_equal(vars[[2]], quote(y))
+  expect_equal(op_sort(out), list(quote(x), quote(y)))
+})
+
+test_that("preserved across compute and collapse", {
+  df1 <- memdb_frame(x = sample(10)) %>% arrange(x)
+
+  df2 <- compute(df1)
+  expect_equal(op_sort(df2), list(quote(x)))
+
+  df3 <- collapse(df1)
+  expect_equal(op_sort(df3), list(quote(x)))
 })