+++ NEWS.md
@@ -1,6 +1,10 @@
 # dplyr 0.5.0.9000
 
-* Database tables now display how they are sorted (#2287)
+* `compute()` and `collapse()` now preserve the "ordering" of rows.
+  This only affects the computation of window functions, as the rest
+  of SQL does not care about row order (#2281).
+
+* Database tables now display how they are ordered (#2287)
 
 * [API] `op_vars()` now returns a list of quoted expressions. This
   enables escaping to happen at the correct time (i.e. when the connection